+++ src/main/java/org/prgrms/springbootbasic/processor/Processor.java
@@ -1,6 +1,6 @@
 package org.prgrms.springbootbasic.processor;
 
-import org.prgrms.springbootbasic.message.Response;
+import org.prgrms.springbootbasic.dto.Response;
 
 public interface Processor {
     Response process();
